raise error instead of assert

+++ sky/provision/aws/instance.py
@@ -168,5 +168,7 @@ def cleanup_ports(
     sg_name = (f'sky-sg-{common_utils.user_and_hostname_hash()}'
                f'-{common_utils.truncate_and_hash_cluster_name(cluster_name)}')
     sgs = ec2.security_groups.filter(GroupNames=[sg_name])
-    assert len(list(sgs)) == 1
+    if len(list(sgs)) != 1:
+        raise ValueError(f'Expected security group {sg_name} not found. '
+                         'Cannot cleanup ports.')
     list(sgs)[0].delete()
